Modified the build server so that the number of simultaneous builds is limited. The default limit is 10, but this can be overridden with the command line option --maxSimultaneousBuilds. When the build server is at maximum capacity, it will respond immediately to any additional build-all-from-zip-async requests with response code 503 (service unavailable). According to rfc2616, section 10, 503 means "The server is currently unable to handle the request due to a temporary overloading or maintenance of the server. The implication is that this is a temporary condition which will be alleviated after some delay."
Added NonQueuingExecutor.java and NonQueuingExecutorTest.java

Modified YoungAndroidProjectService (appengine server code) so that if the
response code for a build-all-from-zip-async request is not 200 (OK), it will
include the response code in the RpcResult that it sends back to the client
code.

Modified BuildCommand (appengine client code) so that if the RpcResult from the
server contains the response code 503, it shows the message "The build server
is currently busy. Please try again in a few minutes." in yellow at the top of
the page.

Modified WaitForBuildResultCommand (appengine client code) so that it doesn't
print a different date/time each time it updates the debugger messages.

Other changes to BuildServer.java:
The command line options are now processed in the main method.
Added --maxSimultaneousBuilds command line option.
Added support for two new requests: buildserver/health and buildserver/vars.
buildserver/health simply returns "ok" if the build server is alive.
buildserver/vars returns a list of helpful values related to runtime, OS,
memory, build requests, and build tasks.