Solved issue 77: Investigate why we might get KeyNotFound exceptions in the server Fixed issue 96: Code should not continue after failing to decrypt user/project id
In Security.java, changed decryptUserId and decryptProjectId to throw
EncryptionException instead of ignoring it and returning bogus values.
Updated ReceiveBuildServlet and SecurityTest.

In JavaWebStart.java:
Added inner class DecryptedIds.
Modified getWebStartProjectPath to use the mercurial build version as a
prefix.
Added method decryptWebStartProjectPath that decrypts a project path and
checks whether the version included in the project path matches the
current Mercurial build version.

In WebStartFileServlet doGet and doPost, if it catches
EncryptionException while decrypting the project path, reply with SC_GONE.
In doPost (not doGet where we have to return the requested file's
contents), if the post succeeds, but the version from the project path
does not match the current Mercurial build version, reply with
SC_CONFLICT.

In YaCodeblocksController, for all cases where a request is sent to the
appengine server, check the response code. If it's either HTTP_GONE and
HTTP_CONFLICT, show a message to the user that says that the app engine
server has been updated and they need to refresh the designer and restart
the blocks editor. Updated YaHttpServerMainTest.

In YaHttpServerMain, if downloading the phone app failed, warn the user
that this error can occur if they've tried to start the blocks editor with
a previously downloaded .jnlp file.

In FeedbackReporter.showErrorMessageWithExit, removed the trailing " + "
at the end of the message.

Updated the appengine build so that it doesn't rebuild AiSharedLib if it
is already up-to-date.