This change addresses issues 86 (https://code.google.com/p/app-inventor/issues/detail?id=86) about setting the file name for blobs and 91 (https://code.google.com/p/app-inventor/issues/detail?id=91) about deleting orphaned blobs when we delete a project or fail during project creation.
When we have blobs that we know we won't need (e.g., a project
creation that needs to be retried or a project that is deleted) we
need to collect up the blob keys and then delete the blobs outside of
any transaction that is running on the datastore. It seems that there
is no way to use the blobstore file API without having it default to
using the existing transactions, and that leads to errors about
transactions on multiple entity groups.

As part of this change I also set the name of a created blob to
include its project id and file name. We don't use the blob name in
the code, but it does help when browsing the datastore and blobstore
from the admin console.