This change list includes several fixes to help us either avoid or better diagnose intermittent server errors related to blobs. It also updates the App Engine libraries to 1.6.2.1 (which may also help with the errors but we're not sure about that---it seems like a good idea in any case). In particular:
- Make sure *all* blob key lookups happen outside of transactions. A previous
fix did not handle the blob reads in ObjectifyStorageIo.exportProjectSourceZip.

- Log an error if we get a RuntimeException while trying to export source zip.
Hopefully we won't be getting these as much, or at all, once the blob read is
not causing it.

- Use the Blobstore API to read bytes instead of going through FileService and
Channels. I'm hoping this will get rid of some of the timeouts.

- Update the App Engine libraries to 1.6.2.1. We know that there were bug fixes
in the latest version and we were on a really old version (1.5.3) so it seems
like a good idea to update, in case that is contributing to any of our problems.